requirements.txt
pandas
tqdm
networkx
imagehash
requests
pillow
beautifulsoup4